backend/requirements.txt
fastapi[standard]
httpx
numpy<2
pillow
python-multipart
#tflite # runtime untuk tensorflow lite
ai-edge-litert